+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineGetPayloadBodiesByHashV2.java
@@ -80,13 +80,16 @@ public class EngineGetPayloadBodiesByHashV2 extends AbstractEngineGetPayloadBodi
 
     final Blockchain blockchain = protocolContext.getBlockchain();
 
-    final List<Optional<String>> blockAccessLists =
+    final List<BlockBodyWithAccessList> results =
         Arrays.stream(blockHashes)
-            .map(blockHash -> getBlockAccessList(blockchain, blockHash))
+            .parallel()
+            .map(hash -> getBlockBodyWithAccessList(blockchain, hash))
             .collect(Collectors.toList());
 
     final List<Optional<BlockBody>> blockBodies =
-        Arrays.stream(blockHashes).map(blockchain::getBlockBody).collect(Collectors.toList());
+        results.stream().map(BlockBodyWithAccessList::body).collect(Collectors.toList());
+    final List<Optional<String>> blockAccessLists =
+        results.stream().map(BlockBodyWithAccessList::accessList).collect(Collectors.toList());
 
     final EngineGetPayloadBodiesResultV2 engineGetPayloadBodiesResultV2 =
         blockResultFactory.payloadBodiesCompleteV2(blockBodies, blockAccessLists);