+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineGetPayloadBodiesByRangeV1.java
@@ -106,6 +106,7 @@ public class EngineGetPayloadBodiesByRangeV1 extends AbstractEngineGetPayloadBod
     EngineGetPayloadBodiesResultV1 engineGetPayloadBodiesResultV1 =
         blockResultFactory.payloadBodiesCompleteV1(
             LongStream.range(startBlockNumber, endExclusiveBlockNumber)
+                .parallel()
                 .mapToObj(
                     blockNumber ->
                         blockchain